Batch run of integration tests
Implemented SetUp and TearDown functionality in a way to allow us to run tests in a batch (all at once). Once it logs in, every consequent test simply starts from the Main Menu and goes further, instead of trying to re-login for each one.

Note: Renamed the class to more properly describe the tests included.